+++ react/javascript/src/getFeatureStatus.ts
@@ -3,13 +3,14 @@ import IGherkinDocument = messages.GherkinDocument
 
 import { Query as CucumberQuery } from '@cucumber/query'
 import { Query as GherkinQuery } from '@cucumber/gherkin-utils'
+import { getWorstTestStepResult } from '@cucumber/messages'
 
 export default function getFeatureStatus(
   document: IGherkinDocument,
   cucumberQuery: CucumberQuery,
   gherkinQuery: GherkinQuery
 ): messages.TestStepResultStatus {
-  return cucumberQuery.getWorstTestStepResult(
+  return getWorstTestStepResult(
     cucumberQuery.getPickleTestStepResults(gherkinQuery.getPickleIds(document.uri))
   ).status
 }